+++ contao/languages/en/modules.php
@@ -1,6 +1,0 @@
-<?php
-
-/**
- * Backend modules
- */
-$GLOBALS['TL_LANG']['MOD']['cleaner'] = ['Cleaner', ''];