+++ translations/contao_tl_settings.de.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'tl_settings' => [
+        'cleanerOrder' => ['Reihenfolge', 'Wenn die Reihenfolge der Ausführung der Cleaner-Jobs relevant ist, können Sie diese hier festlegen.'],
+        'cleaner_legend' => 'Cleaner',
+    ],
+];