+++ translations/contao_tl_settings.en.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'tl_settings' => [
+        'cleanerOrder' => ['Order', 'Here you can define the order of the cleaner.'],
+        'cleaner_legend' => 'Cleaner',
+    ],
+];